# server/docs/openapi/server.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/internal/observability/pipeline": {
      "get": {
        "description": "Returns the last N agentic_jobs rows with their stage, status, result, and the matching memory + project IDs. Useful for watching the pipeline live in dev.",
        "operationId": "ObservabilityController_pipeline",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "project_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memory_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Recent worker pipeline activity (agentic_jobs + outbox + brief stats)",
        "tags": [
          "Internal Observability"
        ]
      }
    },
    "/v1/internal/observability/memories/{id}/timeline": {
      "get": {
        "operationId": "ObservabilityController_memoryTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Per-memory timeline: pipeline jobs + recall accesses for one memory",
        "tags": [
          "Internal Observability"
        ]
      }
    },
...
    "/v1/observability/summary": {
      "get": {
        "operationId": "ObservabilityController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-bearer": []
          }
        ],
        "tags": [
          "Observability"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenKT API",
    "description": "Canonical NestJS backend contract for OpenKT. Success responses use `{ data, error, meta }` envelopes and JWT-authenticated routes expect `Authorization: Bearer <token>`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token for user-authenticated routes."
      }
    },
...
  }
}